requirements.txt
flask==2.0.1
flask_ngrok==0.0.25
opencv-python==4.5.2.52
pathlib2==2.3.6
pillow==8.3.1
urllib3==1.26.6